+++ client/public/locales/de/translation.json
@@ -48,10 +48,8 @@
   "apply_changes": "Änderungen übernehmen",
   "apply_existing_account_amount_to_goal": "Den vorhandenen Kontostand auf das Ziel anrechnen?",
   "asset": "Vermögen",
-  "asset_deleted_successfully_message": "Vermögenswert erfolgreich gelöscht.",
   "asset_details": "Vermögenswertdetails",
   "asset_name": "Name des Vermögenswerts",
-  "asset_restored_successfully_message": "Vermögenswert erfolgreich wiederhergestellt.",
   "asset_type": "Vermögensart",
   "asset_type_name": "Vermögensartname",
   "asset_types": "Vermögensarten",
