+++ client/public/locales/en-us/translation.json
@@ -50,10 +50,8 @@
   "apply_changes": "Apply Changes",
   "apply_existing_account_amount_to_goal": "Apply existing account balance towards goal target?",
   "asset": "Asset",
-  "asset_deleted_successfully_message": "Asset deleted successfully.",
   "asset_details": "Asset Details",
   "asset_name": "Asset Name",
-  "asset_restored_successfully_message": "Asset restored successfully.",
   "asset_type": "Asset Type",
   "asset_type_name": "Asset Type Name",
   "asset_types": "Asset Types",
