+++ client/public/locales/fr/translation.json
@@ -48,10 +48,8 @@
   "apply_changes": "Appliquer les modifications",
   "apply_existing_account_amount_to_goal": "Appliquer le solde existant du compte à l’objectif cible ?",
   "asset": "Actif",
-  "asset_deleted_successfully_message": "Actif supprimé avec succès.",
   "asset_details": "Détails de l’actif",
   "asset_name": "Nom de l’actif",
-  "asset_restored_successfully_message": "Actif restauré avec succès.",
   "asset_type": "Type d'actif",
   "asset_type_name": "Nom de type d'actif",
   "asset_types": "Types d'actif",
