+++ client/public/locales/it/translation.json
@@ -48,10 +48,8 @@
   "apply_changes": "Applica le modifiche",
   "apply_existing_account_amount_to_goal": "Utilizzare il saldo attuale del conto per raggiungere l'obiettivo?",
   "asset": "Risorsa",
-  "asset_deleted_successfully_message": "Asset eliminato con successo.",
   "asset_details": "Dettagli asset.",
   "asset_name": "Nome dell' asset.",
-  "asset_restored_successfully_message": "Asset recuperato con successo.",
   "asset_type": "Tipo di risorsa",
   "asset_type_name": "Nome del tipo di risorsa",
   "asset_types": "Tipi di risorse",
