+++ client/public/locales/zh-hans/translation.json
@@ -45,10 +45,8 @@
   "apply_changes": "应用变更",
   "apply_existing_account_amount_to_goal": "将现有账户余额计入目标金额？",
   "asset": "资产",
-  "asset_deleted_successfully_message": "资产删除成功。",
   "asset_details": "资产详情",
   "asset_name": "资产名称",
-  "asset_restored_successfully_message": "资产恢复成功。",
   "assets": "资产",
   "assets_settings": "资产设置",
   "authorization_code_missing_message": "授权码缺失，请检查回调 URL 是否正确。",
