+++ client/src/hooks/mutations/assets/useRestoreAssetsMutation.ts
@@ -1,0 +1,27 @@
+import { notifications } from "@mantine/notifications";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { AxiosError } from "axios";
+import { assetsQueryKey, translateAxiosError } from "~/helpers/requests";
+import { useAuth } from "~/providers/AuthProvider/AuthProvider";
+
+export const useRestoreAssetsMutation = () => {
+  const queryClient = useQueryClient();
+  const { request } = useAuth();
+
+  return useMutation({
+    mutationFn: async (assetId: string) =>
+      await request({
+        url: `/api/asset/restore`,
+        method: "POST",
+        params: { assetId },
+      }),
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: [assetsQueryKey] });
+    },
+    onError: (error: AxiosError) =>
+      notifications.show({
+        color: "var(--button-color-destructive)",
+        message: translateAxiosError(error),
+      }),
+  });
+};
